requirements.txt
pygame==2.4.0
python-chess==1.999
